// src/types/preknow.ts
export interface Product {
  _id: string;
  name: string;
  author: string;
  manufacturer: string;
  description: string;
  price: number;
  oldPrice: number;
  imageUrl: string;
  shop: Shop;
  category: Category;
  category_slug: string;
  quantity: number;
  in_wishlist: boolean;
  variations: any;
  created_at: string;
  updated_at: string;
  numberOfPage: number;
  manufacture_at: string;
  dimension: string;
}

export interface RatingCount {
  rating: number;
  total: number;
}

export interface Category {
  _id: string;
  name: string;
  slug: string;
  details: string;
  image: string;
}

export interface Banner {
  id: string;
  title: string;
  description: string;
  image: Attachment;
}

export interface Type {
  id: string;
  name: string;
  slug?: any;
  banners: Banner[];
  icon?: string;
  promotional_sliders: Attachment[];
  settings: {
    isHome: boolean;
    layoutType: string;
    productCard?: string;
  };
}

export interface Shop {
  _id: string;
  name: string;
  description: string;
}